Rename plot time/frequency dialog
